requirements.txt
click==8.3.2
pytaglib>=2.1.0,<3
PyMySQL==1.1.2
tqdm==4.67.3
